Backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
python-decouple
gunicorn
whitenoise
psycopg2-binary
Pillow
